+++ src/sgl/device/resource.cpp
@@ -665,6 +665,20 @@ TextureView::TextureView(ref<Device> device, ref<Texture> texture, TextureViewDe
     );
 }
 
+DescriptorHandle TextureView::descriptor_handle_ro() const
+{
+    rhi::DescriptorHandle rhi_handle = {};
+    m_rhi_texture_view->getDescriptorHandle(rhi::DescriptorHandleAccess::Read, &rhi_handle);
+    return DescriptorHandle(rhi_handle);
+}
+
+DescriptorHandle TextureView::descriptor_handle_rw() const
+{
+    rhi::DescriptorHandle rhi_handle = {};
+    m_rhi_texture_view->getDescriptorHandle(rhi::DescriptorHandleAccess::ReadWrite, &rhi_handle);
+    return DescriptorHandle(rhi_handle);
+}
+
 NativeHandle TextureView::native_handle() const
 {
     rhi::NativeHandle rhi_handle = {};