+++ src/sgl/device/resource.h
@@ -620,6 +620,9 @@ public:
     const SubresourceRange& subresource_range() const { return m_desc.subresource_range; }
     std::string_view label() const { return m_desc.label; }
 
+    DescriptorHandle descriptor_handle_ro() const;
+    DescriptorHandle descriptor_handle_rw() const;
+
     /// Get the native texture view handle.
     NativeHandle native_handle() const;