+++ tests/slangpy/device/test_print.py
@@ -13,6 +13,8 @@ import sglhelpers as helpers
 def test_print(device_type: spy.DeviceType):
     if device_type == spy.DeviceType.metal:
         pytest.skip("Slang bug https://github.com/shader-slang/slang/issues/6764")
+    if device_type == spy.DeviceType.cuda:
+        pytest.skip("Slang crash")
 
     device = spy.Device(type=device_type, enable_print=True)
     helpers.dispatch_compute(